+++ src/internal.rs
@@ -1,1 +1,2 @@
+pub(crate) mod fuzzy;
 pub(crate) mod list;